requirements.txt
Flask==2.3.3
pandas==1.5.3
numpy==1.24.3
openpyxl==3.1.2
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0